chore!: rename package prove to proof
