+++ pkg/proof/proof_test.go
@@ -1,4 +1,4 @@
-package prove
+package proof
 
 import (
 	"sort"
